Add untested CostBasedLru, which is an LRU cache that evicts based on item cost, not count.
Tests and fixes are coming next.

+++ src/estimate_cost.rs
@@ -3,5 +3,5 @@
 /// The caches in this crate will cache up to a specified total cost, then begin
 /// evicting entries which are least recently used.
 pub trait EstimateCost {
-    fn estimate_cost(&self) -> usize;
+    fn estimate_cost(&self) -> u64;
 }
